refactor: member Model 클래스 위치 변
