+++ src/main/java/org/c4marathon/assignment/domain/repository/transfertransaction/TransferTransactionQueryRepository.java
@@ -4,7 +4,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Map;
 
-import org.c4marathon.assignment.domain.model.Member;
+import org.c4marathon.assignment.domain.model.member.Member;
 import org.c4marathon.assignment.domain.model.transfer.TransferTransaction;
 import org.springframework.stereotype.Repository;
