+++ src/main/java/org/c4marathon/assignment/domain/repository/transfertransaction/TransferTransactionQueryRepositoryImpl.java
@@ -6,10 +6,10 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import org.c4marathon.assignment.domain.model.Member;
-import org.c4marathon.assignment.domain.model.QMember;
+import org.c4marathon.assignment.domain.model.member.Member;
 import org.c4marathon.assignment.domain.model.account.QMainAccount;
 import org.c4marathon.assignment.domain.model.enums.TransferStatus;
+import org.c4marathon.assignment.domain.model.member.QMember;
 import org.c4marathon.assignment.domain.model.transfer.QTransferTransaction;
 import org.c4marathon.assignment.domain.model.transfer.TransferTransaction;
 import org.springframework.stereotype.Repository;
