+++ src/main/java/org/c4marathon/assignment/domain/service/MemberService.java
@@ -2,7 +2,7 @@ package org.c4marathon.assignment.domain.service;
 
 import lombok.RequiredArgsConstructor;
 
-import org.c4marathon.assignment.domain.model.Member;
+import org.c4marathon.assignment.domain.model.member.Member;
 import org.c4marathon.assignment.dto.member.MemberRegistrationRequestDto;
 import org.c4marathon.assignment.domain.repository.MemberRepository;
 import org.springframework.stereotype.Service;
