+++ src/main/java/org/c4marathon/assignment/domain/service/TransferTransactionService.java
@@ -6,7 +6,7 @@ import java.util.List;
 import java.util.Map;
 
 import org.c4marathon.assignment.infra.config.property.TransferTransactionPolicyProperties;
-import org.c4marathon.assignment.domain.model.Member;
+import org.c4marathon.assignment.domain.model.member.Member;
 import org.c4marathon.assignment.domain.model.account.MainAccount;
 import org.c4marathon.assignment.domain.model.transfer.TransferTransaction;
 import org.c4marathon.assignment.domain.repository.mainaccount.MainAccountRepository;
