+++ src/main/java/org/c4marathon/assignment/usecase/MemberAccountUseCase.java
@@ -1,6 +1,6 @@
 package org.c4marathon.assignment.usecase;
 
-import org.c4marathon.assignment.domain.model.Member;
+import org.c4marathon.assignment.domain.model.member.Member;
 import org.c4marathon.assignment.dto.member.MemberRegistrationRequestDto;
 import org.c4marathon.assignment.domain.service.mainaccount.MainAccountService;
 import org.c4marathon.assignment.domain.service.MemberService;
